필터 리스트 뷰 디자인 수정
